Clean up some compiler warnings on (un)signed comparisons, initialisation order

+++ src/games/gameexpl.cc
@@ -108,7 +108,7 @@ Array<int> GameExplicitRep::NumStrategies() const
 {
   const_cast<GameExplicitRep *>(this)->BuildComputedValues();
   Array<int> dim(m_players.size());
-  for (int pl = 1; pl <= m_players.size(); pl++) {
+  for (size_t pl = 1; pl <= m_players.size(); pl++) {
     dim[pl] = m_players[pl]->m_strategies.size();
   }
   return dim;
